+++ src/Communication/SendQueue.h
@@ -7,7 +7,7 @@ class SendQueue {
  public:
     static void addToQueue(Frame dataToSend);
     static uint8_t getItemToSend();
-    static int count;
+    static uint8_t count;
 
  private:
     static uint8_t itemsInQueue[20];
